Reboot if the Linux kernel panics
Prior to this commit, Linux would halt indefinitely. If enabled and
started, a hardware watchdog would reboot in theory, but it's possible
to panic in areas unguarded by the AM335x watchdog. Since no good can
come from a panic, pause for a little bit (10 seconds) and reboot. The
pause is good for throttling the continuous panic case, but the hope
would be that if this is an on-boot panic, u-boot reverts to the
previously running firmware.